requirements.txt
django>=5.0,<7
python-dotenv>=1.0
requests>=2.31
gunicorn>=23.0
whitenoise>=6.8
